// src/api/generated/model/flexAggregateDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */

export interface FlexAggregateDto {
  /**
   * Timestamp of the aggregated reading
   * @nullable
   */
  time: string | null;
  /** Aggregated grid return power for the group */
  return: number;
  /** Aggregated grid delivery power for the group */
  delivery: number;
  /** Available steerable power to maintain zero export (in watts) for the group */
  steerablePowerZeroExport: number;
  /** Total charger consumption power for the group */
  chargerConsumption: number;
  /** Total solar inverter production power for the group */
  solarProduction: number;
  /** Number of smart meters contributing to this aggregate */
  smartMeterCount: number;
  /** Number of solar inverters contributing to this aggregate */
  solarInverterCount: number;
  /** Number of chargers contributing to this aggregate */
  chargerCount: number;
}